renames files and finished button variants
