+++ src/components/Button/Button.scss
@@ -1,16 +1,29 @@
 .button {
-  padding: 5px 15px;
-  border: none;
+  padding: 3px 13px;
   border-radius: 4px;
   font-size: 15px;
+  box-sizing: border-box;
+  border: 1px solid transparent;
 
-  &.primary {
+  &:hover {
+    transition-duration: 200ms;
+  }
+
+  &.dark {
+    border-color: #8b92c0;
     background-color: #25283d;
     color: #ffff;
+    &:hover {
+      background-color: #1e1f31;
+    }
+  }
+
+  &.light {
+    background-color: #ffff;
+    border-color: #25283d;
 
     &:hover {
-      background-color: #202232;
-      transition-duration: 200ms;
+      background-color: #f4f1f8;
     }
   }
 }
