+++ src/components/deck-cover/deck-cover.scss
@@ -25,21 +25,26 @@
     display: flex;
     flex-direction: column;
     background-color: #efefef;
+
     label {
       font-size: 20px;
     }
+
     p {
       font-size: 12px;
       flex: 1;
     }
+
     .button-strip {
       display: flex;
+
       div {
         flex: 1;
       }
 
       button {
         margin-left: 15px;
+        border-color: transparent;
       }
     }
   }
