+++ src/components/flip-tile/flip-tile.scss
