+++ src/models/card-content.ts
